deployment/requirements.txt
streamlit
pandas
seaborn
matplotlib
numpy
plotly
Pillow
wordcloud
regex
nltk
scikit-learn==1.0.1
feature-engine==1.5.2
tensorflow-cpu==2.12.0